# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Marty gRPC Service API
  description: |
    API documentation for the Marty gRPC service ecosystem for passport issuance and verification.
    This documentation complements the gRPC service definitions and provides a RESTful interface
    for easier testing and integration.
  version: 1.0.0
  contact:
    name: Marty Development Team
    email: support@marty.example.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.marty.example.com
    description: Production server
...
paths:
...
  /v1/pkd/masterlist:
    get:
      tags:
        - PKD
      summary: Download CSCA Master List
      description: Retrieve the latest CSCA Master List
      operationId: getMasterList
      parameters:
        - name: country
          in: query
          description: Filter master list by country code (ISO 3166-1 alpha-3)
          required: false
          schema:
            type: string
            example: USA
      responses:
        '200':
          description: Successful operation
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
                description: ASN.1 encoded master list
            application/json:
              schema:
                $ref: '#/components/schemas/MasterListResponse'
    post:
      tags:
        - PKD
      summary: Upload CSCA Master List
      description: Upload or update a CSCA Master List
      operationId: uploadMasterList
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
              description: ASN.1 encoded master list
      responses:
        '201':
          description: Master list uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MasterListUploadResponse'
  /v1/pkd/masterlist/{country}:
    get:
      tags:
        - PKD
      summary: Get country-specific master list
      description: Retrieve the CSCA Master List for a specific country
      operationId: getCountryMasterList
      parameters:
        - name: country
          in: path
          description: Country code (ISO 3166-1 alpha-3)
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
                description: ASN.1 encoded master list
            application/json:
              schema:
                $ref: '#/components/schemas/MasterListResponse'
  /v1/pkd/dsclist:
    get:
      tags:
        - PKD
      summary: Download Document Signer Certificate List
      description: Retrieve the latest DSC list
      operationId: getDscList
      parameters:
        - name: country
          in: query
          description: Filter DSC list by country code (ISO 3166-1 alpha-3)
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
                description: ASN.1 encoded DSC list
            application/json:
              schema:
                $ref: '#/components/schemas/DscListResponse'
    post:
      tags:
        - PKD
      summary: Upload Document Signer Certificate List
      description: Upload or update a Document Signer Certificate List
      operationId: uploadDscList
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
              description: ASN.1 encoded DSC list
      responses:
        '201':
          description: DSC list uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DscListUploadResponse'
  /v1/pkd/dsclist/{country}:
    get:
      tags:
        - PKD
      summary: Get country-specific DSC list
      description: Retrieve Document Signer Certificates for a specific country
      operationId: getCountryDscList
      parameters:
        - name: country
          in: path
          description: Country code (ISO 3166-1 alpha-3)
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
                description: ASN.1 encoded DSC list
            application/json:
              schema:
                $ref: '#/components/schemas/DscListResponse'
  /v1/pkd/crl:
    get:
      tags:
        - PKD
      summary: Download Certificate Revocation Lists
      description: Retrieve the latest Certificate Revocation Lists
      operationId: getCrlList
      parameters:
        - name: country
          in: query
          description: Filter CRL by country code (ISO 3166-1 alpha-3)
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
                description: ASN.1 encoded CRL
            application/json:
              schema:
                $ref: '#/components/schemas/CrlResponse'
    post:
      tags:
        - PKD
      summary: Upload Certificate Revocation List
      description: Upload or update a Certificate Revocation List
      operationId: uploadCrl
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
              description: ASN.1 encoded CRL
      responses:
        '201':
          description: CRL uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrlUploadResponse'
  /v1/pkd/crl/{country}:
    get:
      tags:
        - PKD
      summary: Get country-specific CRL
      description: Retrieve the Certificate Revocation List for a specific country
      operationId: getCountryCrl
      parameters:
        - name: country
          in: path
          description: Country code (ISO 3166-1 alpha-3)
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
                description: ASN.1 encoded CRL
            application/json:
              schema:
                $ref: '#/components/schemas/CrlResponse'
...
components:
  schemas:
    Certificate:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        subject:
          type: string
          example: "CN=CSCA-USA,O=Department of State,C=US"
        issuer:
          type: string
          example: "CN=CSCA-USA,O=Department of State,C=US"
        validFrom:
          type: string
          format: date-time
        validTo:
          type: string
          format: date-time
        publicKey:
          type: string
        certificateData:
          type: string
          format: byte
          description: Base64 encoded X.509 certificate
        status:
          type: string
          enum:
            - ACTIVE
            - EXPIRED
            - REVOKED
      required:
        - id
        - subject
        - issuer
        - validFrom
        - validTo
        - certificateData
        - status
...
    MasterListResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        version:
          type: integer
          example: 42
        created:
          type: string
          format: date-time
        countries:
          type: array
          items:
            type: string
            example: "USA"
        certificates:
          type: array
          items:
            $ref: '#/components/schemas/Certificate'
      required:
        - id
        - version
        - created
        - countries
        - certificates
    MasterListUploadResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        version:
          type: integer
        created:
          type: string
          format: date-time
        status:
          type: string
          enum:
            - PROCESSED
            - PENDING
            - ERROR
        certificateCount:
          type: integer
      required:
        - id
        - version
        - created
        - status
        - certificateCount
    DscListResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        version:
          type: integer
        created:
          type: string
          format: date-time
        countries:
          type: array
          items:
            type: string
        certificates:
          type: array
          items:
            $ref: '#/components/schemas/Certificate'
      required:
        - id
        - version
        - created
        - countries
        - certificates
    DscListUploadResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        version:
          type: integer
        created:
          type: string
          format: date-time
        status:
          type: string
          enum:
            - PROCESSED
            - PENDING
            - ERROR
        certificateCount:
          type: integer
      required:
        - id
        - version
        - created
        - status
        - certificateCount
    CrlResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        version:
          type: integer
        created:
          type: string
          format: date-time
        issuer:
          type: string
        thisUpdate:
          type: string
          format: date-time
        nextUpdate:
          type: string
          format: date-time
        revokedCertificates:
          type: array
          items:
            $ref: '#/components/schemas/RevokedCertificate'
      required:
        - id
        - version
        - created
        - issuer
        - thisUpdate
        - nextUpdate
    CrlUploadResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        version:
          type: integer
        created:
          type: string
          format: date-time
        status:
          type: string
          enum:
            - PROCESSED
            - PENDING
            - ERROR
        revokedCount:
          type: integer
      required:
        - id
        - version
        - created
        - status
        - revokedCount
    RevokedCertificate:
      type: object
      properties:
        serialNumber:
          type: string
        revocationDate:
          type: string
          format: date-time
        reasonCode:
          type: integer
      required:
        - serialNumber
        - revocationDate
...